Updated Road Image + Working on fuel costs + Bug Fix
Updated road art so that instead of a dotted white line it is now a straight line. Moved old image to 'OLD' asset folder
Roads not showing up has been fixed, this was due to how it was arranged in unity editor, parent object under which the roads were stored had its Z value changed to '-2000' which caused all the roads to have the same Z value.
Work has been started on implementing fuel costs with a new function CalcFeulCost in PlayerMovement.cs. A change has been made in MoneyDisplayChange.cs to accomodate a negative value change.
BUG
 - When moving between cities in multiple locations no change is made to the money, this has been seen as NaN or as 0 money being calculated, it has something to do in CalcFeulCost, but cannot find the reason yet.
Also this is the final Unity commit for the night.