Escape menu + Connecting item menus with the database + cleanup
Made the escape menu, with full functionality

clean up of uneeded controls i keyboardcontrols.cs

Finished UIDetection in itemmenu.cs

Added
- FindLocationIDOfCurrentLocation()
- FindProductIDOfInputString()
to itemmenu.cs

Bug encountered in SQL (that I didn't have before I swear) where it considers the value being compared against in the 'WHERE' part to be a column that is being compared instead, or it doesn't recognize the column name for whatever reason.

ResolveSell is almost connected to work fully but due to the bug above it doesnt work

You are now able to see the correct items and values being displayed for all locations.